+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.3.4 - 2020-12-11
+* Revert Pebble to Badger.
+* **Breaking**: bumps network and database versions
+
+# v0.3.3 - 2020-12-10
+* Fix sync issue.
+* Fix pkger issue.
+* **Breaking**: bumps network and database versions
+
 # v0.3.2 - 2020-12-09
 * Switch from BadgerDB to Pebble.
 * Add FPC statements.